editoast: moving PhysicsConsistParameters from editoast/core_client/src/simulation.rs to editoast/src/views/timetable.rs
