+++ editoast/src/views/timetable/stdcm.rs
@@ -14,7 +14,6 @@ use core_client::CoreClient;
 use core_client::conflict_detection::TrainRequirements;
 use core_client::pathfinding::InvalidPathItem;
 use core_client::pathfinding::PathfindingResultSuccess;
-use core_client::simulation::PhysicsConsistParameters;
 use core_client::simulation::RoutingRequirement;
 use core_client::simulation::SpacingRequirement;
 use editoast_authz as authz;
@@ -56,6 +55,7 @@ use crate::views::AuthenticationExt;
 use crate::views::AuthorizationError;
 use crate::views::path::pathfinding::PathfindingResult;
 use crate::views::timetable::Conflict;
+use crate::views::timetable::PhysicsConsistParameters;
 use crate::views::timetable::simulation;
 use crate::views::timetable::simulation::consist_train_simulation_batch;
 use crate::views::timetable::train_simulation_batch;
